requirements.txt
ultralytics==8.0.96
opencv-python~=4.7.0.72
torch~=2.0.0
Pillow~=9.5.0
numpy~=1.24.3